requirements.txt
# Core dependencies
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=4.5.0
fastapi>=0.100.0
uvicorn>=0.22.0
click>=8.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.2
schedule>=1.2.0
pyyaml>=6.0
rich>=13.0.0
python-dateutil>=2.8.0

# Traditional NLP analysis dependencies
spacy>=3.7.0
nltk>=3.8.0
textblob>=0.17.1
scikit-learn>=1.3.0
numpy>=1.24.0
textstat>=0.7.0

# Apache Tika PDF processing (existing dependency used by parsers)
requests-toolbelt>=1.0.0

# Optional dependencies for development
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0